refactor(scripts): rename emit_otel_genai_spans → emit_genai_spans
Fixes #529 — action_target naming convention violation.
Old script retained as deprecation stub for backward compatibility.
All 12 references updated across docs, tests, and README.
