+++ docs/plans/2026-03-26-q2-wave2-phase3-4.md
@@ -359,11 +359,11 @@ outcomes.
   1 if any degraded, 2 if any unreachable (#342)
 - `tests/test_health_check_services.py` — tests healthy, degraded, and unreachable
   scenarios with mocked HTTP (#342)
-- `scripts/emit_otel_genai_spans.py` — emits GenAI semantic convention attributes
+- `scripts/emit_genai_spans.py` — emits GenAI semantic convention attributes
   (`gen_ai.system`, `gen_ai.request.model`, `gen_ai.usage.input_tokens`,
   `gen_ai.usage.output_tokens`, `gen_ai.response.finish_reason`) for every LLM span;
   extends `scripts/instrument_agent_calls.py` (#369)
-- `tests/test_emit_otel_genai_spans.py` — verifies all required GenAI attributes are
+- `tests/test_emit_genai_spans.py` — verifies all required GenAI attributes are
   present in emitted spans (#369)
 - `pyproject.toml` updated: `opentelemetry-sdk` and `opentelemetry-exporter-otlp`
   added to project dependencies
@@ -374,14 +374,14 @@ outcomes.
 
 #### Acceptance Criteria
 1. `uv run pytest tests/test_instrument_agent_calls.py tests/test_health_check_services.py
-   tests/test_emit_otel_genai_spans.py -q` exits 0 — PASS/FAIL
-2. `uv run python scripts/emit_otel_genai_spans.py --dry-run` exits 0 and prints a
+   tests/test_emit_genai_spans.py -q` exits 0 — PASS/FAIL
+2. `uv run python scripts/emit_genai_spans.py --dry-run` exits 0 and prints a
    sample span with all required GenAI attributes — PASS/FAIL
 3. `uv run python scripts/health_check_services.py --check` exits 0 against
    `data/substrate-atlas.yml` (at least 1 service entry must be present) — PASS/FAIL
 4. `opentelemetry-sdk` appears in `pyproject.toml` dependencies — PASS/FAIL
 5. `uv run ruff check scripts/instrument_agent_calls.py scripts/health_check_services.py
-   scripts/emit_otel_genai_spans.py` exits 0 — PASS/FAIL
+   scripts/emit_genai_spans.py` exits 0 — PASS/FAIL
 
 ---
 
@@ -411,7 +411,7 @@ outcomes.
 - `tests/test_measure_verification_effort.py` — tests with fixture scratchpad files
   covering single phase, multiple phases, and missing timestamps (#345)
 - `scripts/detect_slow_error_discovery.py` — reads OTel span data (JSONL from
-  `scripts/emit_otel_genai_spans.py`); flags phases where error discovery latency
+  `scripts/emit_genai_spans.py`); flags phases where error discovery latency
   exceeds threshold from `data/governance-thresholds.yml`; exits 0 ok, 1 slow (#346)
 - `tests/test_detect_slow_error_discovery.py` — tests fast-path and slow-path
   detection against fixture JSONL (#346)
