+++ scripts/instrument_agent_calls.py
@@ -14,7 +14,7 @@ Inputs:
 Outputs:
     - Spans emitted to stdout (ConsoleSpanExporter) or OTLP endpoint
     - Returns tracer instance for use in agent scripts
-        - When combined with scripts/emit_otel_genai_spans.py, span-close hooks append measured
+        - When combined with scripts/emit_genai_spans.py, span-close hooks append measured
             token usage to session_cost_log.json for model calls this repo controls directly
 
 Usage:
